Add files via upload
I'm trying to get the background image placed behind only the menu and header. tried putting it in class in a div but that moves most of the work  around and leaves the menu and header below the image. Also attempted putting it directly into the menu (li) but that doesnt work out well.